Issue #11320:change sDelete to either delete or close
git-svn-id: http://svn.code.sf.net/p/postbooks/code/xtuple/trunk@9467 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ guiclient/purchaseOrder.cpp
@@ -934,21 +934,36 @@ void purchaseOrder::sEdit()
 
 void purchaseOrder::sDelete()
 {
-  if (QMessageBox::question(this, tr("Delete Purchase Order Item?"),
-			                      tr("<p>Are you sure you want to delete this "
-				                     "Purchase Order Line Item?"),
-	  QMessageBox::Yes,
-	  QMessageBox::No | QMessageBox::Default) == QMessageBox::No)
-    return;
-
-  q.prepare( "DELETE FROM poitem "
-             "WHERE (poitem_id=:poitem_id);" );
-  q.bindValue(":poitem_id", _poitem->id());
-  q.exec();
-  if (q.lastError().type() != QSqlError::NoError)
+  if (_deleteMode == cDelete)
   {
-    systemError(this, q.lastError().databaseText(), __FILE__, __LINE__);
-    return;
+    if (QMessageBox::question(this, tr("Delete Purchase Order Item?"),
+                                              tr("<p>Are you sure you want to delete this "
+				                     "Purchase Order Line Item?"),
+           QMessageBox::Yes,
+            QMessageBox::No | QMessageBox::Default) == QMessageBox::No)
+      return;
+
+    q.prepare( "DELETE FROM poitem "
+               "WHERE (poitem_id=:poitem_id);" );
+    q.bindValue(":poitem_id", _poitem->id());
+    q.exec();
+    if (q.lastError().type() != QSqlError::NoError)
+    {
+      systemError(this, q.lastError().databaseText(), __FILE__, __LINE__);
+      return;
+    }
+  }
+  else
+  {
+    q.prepare( "UPDATE poitem SET poitem_status='C' "
+               "WHERE (poitem_id=:poitem_id);" );
+    q.bindValue(":poitem_id", _poitem->id());
+    q.exec();
+    if (q.lastError().type() != QSqlError::NoError)
+    {
+      systemError(this, q.lastError().databaseText(), __FILE__, __LINE__);
+      return;
+    }
   }
 
   sFillList();
